refactor: group markdown entities
