+++ src/features/NoteEditor/RichEditor/plugins/Markdown/nodes/FormattingNode.tsx
